refactor: move appInfoSchema to common package
