+++ packages/backend/src/modules/app-lifecycle/app-lifecycle.service.ts
@@ -3,8 +3,8 @@ import { createAppUrn, extractAppUrn } from '@/common/helpers/app-helpers';
 import { ConfigurationService } from '@/core/config/configuration.service';
 import { LoggerService } from '@/core/logger/logger.service';
 import { SSEService } from '@/core/sse/sse.service';
-import type { AppUrn } from '@/types/app/app.types';
 import { HttpStatus, Injectable } from '@nestjs/common';
+import type { AppUrn } from '@runtipi/common/types';
 import { lt, valid } from 'semver';
 import semver from 'semver';
 import validator, { isFQDN } from 'validator';
