+++ packages/backend/src/modules/app-lifecycle/commands/generate-env-command.ts
@@ -4,7 +4,7 @@ import type { AppHelpers } from '@/modules/apps/app.helpers';
 import type { DockerService } from '@/modules/docker/docker.service';
 import { MarketplaceService } from '@/modules/marketplace/marketplace.service';
 import type { AppEventFormInput } from '@/modules/queue/entities/app-events';
-import type { AppUrn } from '@/types/app/app.types';
+import type { AppUrn } from '@runtipi/common/types';
 import { AppLifecycleCommand } from './command';
 
 export class GenerateAppEnvCommand extends AppLifecycleCommand {
