+++ packages/backend/src/modules/app-lifecycle/commands/reset-app-command.ts
@@ -5,7 +5,7 @@ import { DockerService } from '@/modules/docker/docker.service';
 import type { EnvUtils } from '@/modules/env/env.utils';
 import { MarketplaceService } from '@/modules/marketplace/marketplace.service';
 import type { AppEventFormInput } from '@/modules/queue/entities/app-events';
-import type { AppUrn } from '@/types/app/app.types';
+import type { AppUrn } from '@runtipi/common/types';
 import { AppLifecycleCommand } from './command';
 
 export class ResetAppCommand extends AppLifecycleCommand {
