+++ packages/backend/src/modules/app-stores/app-store-files-manager.ts
@@ -5,8 +5,8 @@ import type { ConfigurationService } from '@/core/config/configuration.service';
 import type { AppStore } from '@/core/database/drizzle/types';
 import type { FilesystemService } from '@/core/filesystem/filesystem.service';
 import type { LoggerService } from '@/core/logger/logger.service';
-import type { AppUrn } from '@/types/app/app.types';
-import { appInfoSchema } from '../marketplace/dto/marketplace.dto';
+import { appInfoSchema } from '@runtipi/common/schemas';
+import type { AppUrn } from '@runtipi/common/types';
 
 export class AppStoreFilesManager {
   constructor(
