+++ packages/backend/src/modules/apps/__tests__/app.helpers.test.ts
@@ -2,9 +2,9 @@ import { createAppUrn } from '@/common/helpers/app-helpers';
 import { ConfigurationService } from '@/core/config/configuration.service';
 import { FilesystemService } from '@/core/filesystem/filesystem.service';
 import { EnvUtils } from '@/modules/env/env.utils';
-import type { AppInfo } from '@/modules/marketplace/dto/marketplace.dto';
-import type { AppUrn } from '@/types/app/app.types';
 import { Test } from '@nestjs/testing';
+import type { AppInfo } from '@runtipi/common/schemas';
+import type { AppUrn } from '@runtipi/common/types';
 import { fromPartial } from '@total-typescript/shoehorn';
 import { beforeEach, describe, expect, it } from 'vitest';
 import { mock } from 'vitest-mock-extended';
