+++ packages/backend/src/modules/apps/apps.repository.ts
@@ -2,8 +2,8 @@ import { extractAppUrn } from '@/common/helpers/app-helpers';
 import { DATABASE, type Database } from '@/core/database/database.module';
 import { app } from '@/core/database/drizzle/schema';
 import type { AppStatus, NewApp } from '@/core/database/drizzle/types';
-import type { AppUrn } from '@/types/app/app.types';
 import { Inject, Injectable } from '@nestjs/common';
+import type { AppUrn } from '@runtipi/common/types';
 import { and, asc, eq, ne, notInArray } from 'drizzle-orm';
 
 @Injectable()
