+++ packages/backend/src/tests/utils/create-app-in-store.ts
@@ -1,9 +1,9 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { DATA_DIR } from '@/common/constants';
-import type { AppInfo, AppInfoInput } from '@/modules/marketplace/dto/marketplace.dto';
-import type { AppUrn } from '@/types/app/app.types';
 import { faker } from '@faker-js/faker';
+import type { AppInfo, AppInfoInput } from '@runtipi/common/schemas';
+import type { AppUrn } from '@runtipi/common/types';
 
 export const createAppInStore = async (storeId: string, app: Partial<AppInfo> = {}) => {
   const id = app.id ?? faker.string.uuid();
