+++ packages/common/src/schemas/index.ts
@@ -1,6 +1,28 @@
 import { dynamicComposeSchema, serviceSchema } from './dynamic-compose.js';
 import type { DependsOn, Service, ServiceInput } from './dynamic-compose.js';
+
+import { APP_CATEGORIES, ARCHITECTURES, FIELD_TYPES, RANDOM_ENCODINGS, appInfoSchema, formFieldSchema } from './app-info.js';
+import type { AppCategory, AppInfo, AppInfoInput, FieldType, FormField, RandomEncoding } from './app-info.js';
+
 import { toJsonSchema } from './utils/to-json-schema.js';
 
-export { dynamicComposeSchema, serviceSchema, type ServiceInput, type DependsOn, type Service, toJsonSchema };
-
+export {
+  dynamicComposeSchema,
+  serviceSchema,
+  type ServiceInput,
+  type DependsOn,
+  type Service,
+  toJsonSchema,
+  APP_CATEGORIES,
+  formFieldSchema,
+  RANDOM_ENCODINGS,
+  FIELD_TYPES,
+  ARCHITECTURES,
+  appInfoSchema,
+  type AppInfo,
+  type AppInfoInput,
+  type FormField,
+  type FieldType,
+  type RandomEncoding,
+  type AppCategory,
+};
