+++ packages/common/src/types/index.ts
@@ -1,0 +1,3 @@
+import type { AppUrn } from './app-urn.ts';
+
+export type { AppUrn };
